separate GroupInfo and TileConfig
