+++ paddle/cinn/hlir/framework/pir/op_lowering_impl.cc
@@ -38,6 +38,7 @@
 #include "paddle/pir/include/dialect/control_flow/ir/cf_op.h"
 
 #include "paddle/cinn/hlir/dialect/operator/transforms/group_merge/op_with_group_merge_util.h"
+#include "paddle/cinn/ir/group_schedule/config/group_tile_config.h"
 
 PD_DECLARE_bool(cinn_use_cuda_vectorize);
 PD_DECLARE_bool(cinn_enable_map_expr);
@@ -69,175 +70,39 @@ NodeAttr CollectAttrs(const ::pir::Operation& op) {
 
 }  // namespace details
 
-int64_t Next2Power(int64_t n) {
-  if (n == 1) {
-    return 1;
-  }
-  return int64_t(std::pow(2.0, std::ceil(std::log2(n))));
-}
-
-std::shared_ptr<cinn::ir::GroupTileInfo> OpLowererImpl::GetGroupTileInfo(
-    const GroupPtr& group) {
-  std::shared_ptr<cinn::ir::GroupTileInfo> group_tile_info =
-      std::make_shared<cinn::ir::GroupTileInfo>();
-
-  const auto data_dim = group->loop_ranges;
-  group_tile_info->data_rank = data_dim.size();
-  const auto reduce_axis = group->reduce_axis;
-
-  std::set<int64_t> reduce_set;
-  for (auto dim : reduce_axis) {
-    if (dim < 0) {
-      dim += group_tile_info->data_rank;
-    }
-
-    group_tile_info->reduce_axis_.push_back(dim);
-    reduce_set.insert(dim);
-  }
-
-  int64_t spatial_numel = 1;
-  int64_t reduce_numel = 1;
-
-  bool spatial_is_dynamic = false;
-  bool reduce_is_dynamic = false;
-  for (int64_t i = 0; i < group_tile_info->data_rank; ++i) {
-    if (reduce_set.count(i)) {
-      reduce_numel *= data_dim[i];
-      if (data_dim[i] < 0) {
-        reduce_is_dynamic = true;
-      }
-    } else {
-      spatial_numel *= data_dim[i];
-
-      if (data_dim[i] < 0) {
-        spatial_is_dynamic = true;
-      }
-    }
-  }
-
-  bool is_reduce_all =
-      (group_tile_info->reduce_axis_.size() == group_tile_info->data_rank);
-
-  if (is_reduce_all) {
-    reduce_is_dynamic = false;
-  }
-
-  PADDLE_ENFORCE_EQ(
-      reduce_is_dynamic,
-      false,
-      phi::errors::Unimplemented("not support dynamic reduce yet"));
-
-  int64_t reduce_block = 1;
-  int64_t spatial_block = 1;
-
-  int64_t reduce_inner_num = 1;
-  int64_t spatial_inner_num = 1;
-  int warp_num = 1;
-  group_tile_info->is_reduce_all = is_reduce_all;
-
-  if (is_reduce_all) {
-    // warp reduce
-    reduce_block = 1024;
-    spatial_block = 1;
-    spatial_inner_num = 1;
-    reduce_inner_num = 4;
-    warp_num = 8;
-
-  } else if (reduce_numel == 1) {
-    reduce_block = 1;
-    if (spatial_is_dynamic) {
-      spatial_block = 1024;
-
-      reduce_inner_num = 1;
-      warp_num = 8;
-
-      spatial_inner_num = 4;
-
-      group_tile_info->block_num = -1;
-    } else {
-      spatial_block = Next2Power(spatial_numel);
-      if (spatial_block > 1024) {
-        spatial_block = 1024;
-      }
-      reduce_inner_num = 1;
-      warp_num = spatial_block / 128;
-      if (warp_num == 0) {
-        warp_num = 1;
-      }
-      spatial_inner_num = spatial_block / (warp_num * 32);
-      if (spatial_inner_num == 0) {
-        spatial_inner_num = 1;
-      }
-
-      int64_t block_num =
-          int64_t(std::ceil(spatial_numel * 1.0 / spatial_block));
-      group_tile_info->block_num = block_num;
-    }
-  } else if (reduce_numel <= 256) {
-    // warp reduce
-    reduce_block = Next2Power(reduce_numel);
-    spatial_block = 256 / reduce_block;
-    spatial_inner_num = spatial_block;
-    reduce_inner_num = reduce_block / 32;
-    if (reduce_inner_num == 0) {
-      reduce_inner_num = 2;
-    }
-    warp_num = 8;
-  } else if (reduce_numel > 256 && reduce_numel <= 2048) {
-    spatial_block = 1;
-    reduce_block = int64_t(std::ceil(reduce_numel * 1.0 / 256.0)) * 256;
-    warp_num = reduce_block / 256;
-    spatial_inner_num = 1;
-    reduce_inner_num = 8;
-  } else if (reduce_numel > 2048) {
-    spatial_block = 1;
-    reduce_block = 2048;
-    warp_num = 8;
-    reduce_inner_num = int64_t(std::ceil(reduce_numel * 1.0 / 256.0));
-    spatial_inner_num = 1;
-  }
-
-  group_tile_info->reduce_numel = reduce_numel;
-  group_tile_info->reduce_block = reduce_block;
-
-  VLOG(6) << "block num " << group_tile_info->block_num << std::endl;
-  VLOG(6) << "num warp " << warp_num << std::endl;
-  VLOG(6) << "flatten block " << spatial_block << std::endl;
-  VLOG(6) << "reduce block  " << reduce_block << std::endl;
-  VLOG(6) << "flatten inner num " << spatial_inner_num << std::endl;
-  VLOG(6) << "reduce inner num " << reduce_inner_num << std::endl;
-
-  group_tile_info->warp_num = warp_num;
-  group_tile_info->spatial_inner_num = spatial_inner_num;
-  group_tile_info->reduce_inner_num = reduce_inner_num;
-
-  if (reduce_block > 1 && reduce_block <= 256) {
-    group_tile_info->reduce_method = ir::WarpReduceMethod();
-  }
-
+std::shared_ptr<GroupInfo> OpLowererImpl::GetGroupInfo(
+    const GroupPtr& group,
+    const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map) {
+  std::shared_ptr<GroupInfo> group_info = std::make_shared<GroupInfo>();
+  group_info->data_space = group->loop_ranges;
+  group_info->reduce_axis = group->reduce_axis;
   for (auto op : group->ops) {
     if (CompatibleInfo::OpKind(*op) == OpPatternKind::kReduction) {
-      group_tile_info->reduce_tensor_names.insert(ValueName(op->result(0)));
+      group_info->reduce_var_names.insert(ValueName(op->result(0)));
     }
   }
 
-  for (auto& val : group->output_values) {
-    if (val.defining_op()->name() == "cinn_op.reshape" &&
-        erase_reshape.count(val.defining_op())) {
-      group_tile_info->direct_output_var_names.insert(
-          ValueName(val.defining_op()->operand_source(0)));
-    } else {
-      group_tile_info->direct_output_var_names.insert(ValueName(val));
+  BuildBroadcastInfo(group);
+
+  for (auto& op : group->output_ops) {
+    group_info->direct_output_var_names.insert(ValueName(op->result(0)));
+    // collect all output tensor.
+    if (op->name() == "cinn_op.store") {
+      auto input_var_name = ValueName(op->operand_source(0));
+      if (group_info->broadcast_info.count(input_var_name)) {
+        auto base_info = group_info->broadcast_info[input_var_name];
+        base_info.with_constrain = true;
+        group_info->broadcast_info[ValueName(op->result(0))] = base_info;
+      }
+    }
+    for (auto opresult : op->results()) {
+      if (tensor_map.count(opresult) == 0) {
+        continue;
+      }
+      group_info->direct_output_var_names.insert(ValueName(opresult));
     }
   }
-
-  group_tile_info->shared_var_names = shared_var_names;
-  group_tile_info->thread_sync_before_names = thread_sync_before_names;
-
-  group_tile_info->broadcast_info = broadcast_info;
-  group_tile_info->broadcast_to_elementwise = broadcast_to_elementwise;
-
-  return group_tile_info;
+  return group_info;
 }
 
 OpLowererImpl::OpLowererImpl(const Target& target) : target_(target) {
@@ -315,26 +180,7 @@ BucketLoweredFuncsWrapper OpLowererImpl::BucketLower(const GroupPtr& group,
       inner_genevalue.insert(op->result(i));
     }
   }
-
   BuildBroadcastInfo(group);
-
-  for (auto& op : group->output_ops) {
-    // collect all output tensor.
-    if (op->name() == "cinn_op.yield_store") {
-      auto input_var_name = ValueName(op->operand_source(0));
-      if (broadcast_info.count(input_var_name)) {
-        auto base_info = broadcast_info[input_var_name];
-        base_info.with_constrain = true;
-        broadcast_info[ValueName(op->result(0))] = base_info;
-      }
-    }
-
-    for (auto opresult : op->results()) {
-      if (tensor_map.count(opresult) == 0) {
-        continue;
-      }
-    }
-  }
 
   if (apply_group_schedule) {
     std::unordered_set<std::string> output_tensor_names;
@@ -342,14 +188,13 @@ BucketLoweredFuncsWrapper OpLowererImpl::BucketLower(const GroupPtr& group,
       output_tensor_names.insert(ValueName(value));
     }
 
-    std::shared_ptr<cinn::ir::GroupTileInfo> group_tile_info =
-        GetGroupTileInfo(group);
+    std::shared_ptr<GroupInfo> group_info = GetGroupInfo(group, tensor_map);
     std::unique_ptr<ir::GroupScheduler> group_scheduler =
         ir::GroupScheduler::Make(&ir_sch,
                                  output_tensor_names,
                                  target_,
                                  /* is_dy_shape = */ true,
-                                 group_tile_info);
+                                 group_info);
 
     group_scheduler->Schedule();
 
@@ -493,9 +338,9 @@ std::vector<ir::LoweredFunc> OpLowererImpl::LowerMapExpr(
       output_tensor_names.insert(ValueName(value));
     }
 
-    std::shared_ptr<cinn::ir::GroupTileInfo> group_tile_info;
+    std::shared_ptr<hlir::framework::pir::GroupInfo> group_info;
     ir::StaticShapeGroupScheduler group_scheduler(
-        &ir_sch, output_tensor_names, target_, group_tile_info);
+        &ir_sch, output_tensor_names, target_, group_info);
     group_scheduler.MapExprSchedule();
     VLOG(3) << "After group schedule, ir is: \n"
             << ir_sch.GetModule().GetExprs().at(0);
@@ -554,28 +399,7 @@ std::vector<ir::LoweredFunc> OpLowererImpl::LowerGroup(
     }
   }
 
-  BuildBroadcastInfo(group);
-
-  for (auto& op : group->output_ops) {
-    // collect all output tensor.
-    if (op->name() == "cinn_op.yield_store") {
-      auto input_var_name = ValueName(op->operand_source(0));
-      if (broadcast_info.count(input_var_name)) {
-        auto base_info = broadcast_info[input_var_name];
-        base_info.with_constrain = true;
-        broadcast_info[ValueName(op->result(0))] = base_info;
-      }
-    }
-
-    for (auto opresult : op->results()) {
-      if (tensor_map.count(opresult) == 0) {
-        continue;
-      }
-    }
-  }
-
   // 2.Do group schedule.
-
   ir::ModuleExpr mod_expr(func_bodies);
   std::shared_ptr<ir::IRSchedule> ir_sch =
       std::make_shared<ir::IRSchedule>(mod_expr);
@@ -1125,7 +949,7 @@ ir::Expr OpLowererImpl::DoGroupSchedule(
     }
   }
 
-  auto group_tile_info = GetGroupTileInfo(group);
+  std::shared_ptr<GroupInfo> group_info = GetGroupInfo(group, tensor_map);
 
   std::unordered_set<std::string> output_tensor_names;
   for (auto value : group->GetGroupOutputValues()) {
@@ -1136,7 +960,7 @@ ir::Expr OpLowererImpl::DoGroupSchedule(
                                output_tensor_names,
                                target_,
                                /* is_dy_shape = */ true,
-                               group_tile_info);
+                               group_info);
   group_scheduler->Schedule();
   return ir_sch.GetModule().GetExprs().at(0);
 }
