+++ paddle/cinn/hlir/framework/pir/op_lowering_impl.h
@@ -47,6 +47,21 @@ class OpLowererImpl;
 
 typedef bool (OpLowererImpl::*ScheduleDetermineFunction)(::pir::Operation*);
 
+struct GroupInfo {
+  std::vector<int64_t> data_space;
+  std::vector<int64_t> reduce_axis;
+  std::set<std::string> reduce_var_names;
+
+  std::vector<std::string> thread_sync_before_names;
+  std::set<std::string> shared_var_names;
+  std::set<std::string> direct_output_var_names;
+  std::vector<std::string> broadcast_output_names;
+
+  std::unordered_map<std::string, cinn::ir::BroadcastInfo> broadcast_info;
+  std::unordered_map<std::string, cinn::ir::BroadcastInfo>
+      broadcast_to_elementwise;
+};
+
 class OpLowererImpl : public OpLowererImplBase<GroupPtr> {
  public:
   explicit OpLowererImpl(const Target&);
@@ -245,8 +260,9 @@ class OpLowererImpl : public OpLowererImplBase<GroupPtr> {
   ir::Tensor GetTensorSymbolic(const GroupPtr& group,
                                const ::pir::Value& value);
 
-  std::shared_ptr<cinn::ir::GroupTileInfo> GetGroupTileInfo(
-      const GroupPtr& group);
+  std::shared_ptr<GroupInfo> GetGroupInfo(
+      const GroupPtr& group,
+      const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map);
 
   void CollectOutputInfo(::pir::Operation* op,
                          std::vector<Type>* out_types,
@@ -275,12 +291,6 @@ class OpLowererImpl : public OpLowererImplBase<GroupPtr> {
   Target target_;
 
   PrettyNamer* name_gene_;
-
-  std::vector<std::string> thread_sync_before_names;
-  std::set<std::string> shared_var_names;
-  std::set<std::string> direct_output_var_names;
-
-  std::vector<std::string> broadcast_output_names;
 
   std::unordered_map<std::string, cinn::ir::BroadcastInfo> broadcast_info;
   std::unordered_map<std::string, cinn::ir::BroadcastInfo>
