+++ paddle/cinn/ir/group_schedule/base_group_scheduler.cc
@@ -24,13 +24,13 @@ std::unique_ptr<GroupScheduler> GroupScheduler::Make(
     const std::unordered_set<std::string>& output_tensor_names,
     const cinn::common::Target& target,
     bool is_dy_shape,
-    const std::shared_ptr<GroupTileInfo>& group_tile_info) {
+    const std::shared_ptr<hlir::framework::pir::GroupInfo>& group_info) {
   if (is_dy_shape) {
     return std::make_unique<DynamicShapeGroupScheduler>(
-        ir_sch, output_tensor_names, target, group_tile_info);
+        ir_sch, output_tensor_names, target, group_info);
   } else {
     return std::make_unique<StaticShapeGroupScheduler>(
-        ir_sch, output_tensor_names, target, group_tile_info);
+        ir_sch, output_tensor_names, target, group_info);
   }
 }
