+++ paddle/cinn/ir/group_schedule/config/CMakeLists.txt
@@ -1,0 +1,3 @@
+core_gather_headers()
+
+gather_srcs(cinnapi_src SRCS group_tile_config.cc)
