+++ paddle/cinn/ir/group_schedule/dy_shape_group_scheduler.h
@@ -29,8 +29,8 @@ class DynamicShapeGroupScheduler : public GroupScheduler {
       ir::IRSchedule* ir_sch,
       const std::unordered_set<std::string>& output_tensor_names,
       const cinn::common::Target& target,
-      const std::shared_ptr<GroupTileInfo>& group_tile_info)
-      : GroupScheduler(ir_sch, output_tensor_names, target, group_tile_info) {
+      const std::shared_ptr<hlir::framework::pir::GroupInfo>& group_info)
+      : GroupScheduler(ir_sch, output_tensor_names, target, group_info) {
     Init();
   }
